requirements.txt
streamlit
elasticsearch
python-dotenv
expertai-nlapi